refactor: remove unnecessary higher level common folder
