+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/Legend/ReferenceLines.tsx
@@ -14,8 +14,8 @@ import { useProject } from '../../../../hooks/useProject';
 import { type ReferenceLineField } from '../../../common/ReferenceLine';
 import { isCartesianVisualizationConfig } from '../../../LightdashVisualization/VisualizationConfigCartesian';
 import { useVisualizationContext } from '../../../LightdashVisualization/VisualizationProvider';
-import { AddButton } from '../../common/AddButton';
-import { Config } from '../../common/Config';
+import { AddButton } from '../common/AddButton';
+import { Config } from '../common/Config';
 import { ReferenceLine } from './ReferenceLine';
 
 type Props = {
