+++ packages/frontend/src/components/VisualizationConfigs/ChartConfigPanel/common/Config.tsx
